tiktok analytics pythons full structure/my_tiktok_app/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# RAG and AI/ML
langchain==0.0.350
langchain-openai==0.0.2
openai==1.3.7
chromadb==0.4.18
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
numpy==1.24.3

# Vector database
pinecone-client==2.2.4
chromadb==0.4.18

# HTTP and API
requests==2.31.0
httpx==0.25.2

# Data processing
pandas==2.1.3
numpy==1.24.3

# Utilities
python-dateutil==2.8.2
pydantic==2.5.0
pydantic-settings==2.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: For production
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4 